Move file browser lab plugins to a separate file
